requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.4
fastapi-limiter==0.1.6
python-dotenv==1.0.1
pydantic==2.7.1
